+++ src/main/java/com/owncloud/android/ui/fragment/FileDetailSharingFragment.java
@@ -600,7 +600,7 @@ public class FileDetailSharingFragment extends Fragment implements ShareeListAda
         void editExistingShare(OCShare share, int screenTypePermission, boolean isReshareShown,
                                boolean isExpiryDateShown);
 
-        void onLinkShareDownloadLimitFetched(int downloadLimit);
+        void onLinkShareDownloadLimitFetched(long downloadLimit);
 
         void onShareProcessClosed();
     }
